requirements.txt
Flask>=2.3.3
Flask-SocketIO>=5.3.6
python-socketio>=5.8.0
eventlet>=0.33.3
gunicorn>=21.2.0
python-socketio[client]>=5.8.0
Pillow>=10.0.1
Authlib>=1.2.1